form request validations

+++ server-app/app/Http/Services/ProductService.php
@@ -10,10 +10,13 @@ class ProductService{
     public function create($request){
         $product_paths = [];
         try {
-            foreach($request->file('file') as $file){
-                $path =  $file->store('product/'.$request->user()->id, 'public');
-                $product_paths[] = $path;
+            if($request->file('file')){
+                foreach($request->file('file') as $file){
+                    $path =  $file->store('product/'.$request->user()->id, 'public');
+                    $product_paths[] = $path;
+                }
             }
+
             $product = new Product();
             $product->user_id  = $request->user()->id;
             $product->name =  $request->name;
@@ -22,11 +25,14 @@ class ProductService{
             $product->model  = $request->model;
             $product->price = $request->price;
             $product->save();
-            foreach ($product_paths as $path){
-                $product->file()->create([
-                    'path' => $path
-                ]);
+            if($request->file('file')){
+                foreach ($product_paths as $path){
+                    $product->file()->create([
+                        'path' => $path
+                    ]);
+                }
             }
+
             $data = ['status'=>'success','message'=>'Producto Creado Correctamente'];
         } catch (\Throwable $th) {
             Log::error($th);
@@ -40,25 +46,32 @@ class ProductService{
 
     public function update($request){
         try {
-            foreach($request->file('file') as $file){
-                $path =  $file->store('product/'.$request->user()->id, 'public');
-                $product_paths[] = $path;
+            if($request->hasFile('file')){
+                foreach($request->file('file') as $file){
+                    $path =  $file->store('product/'.$request->user()->id, 'public');
+                    $product_paths[] = $path;
+                }
             }
+
             $productUpdate = Product::where('id', $request->id)->with('file')->first();
             $productUpdate->name =  $request->name;
             $productUpdate->description =  $request->description;
             $productUpdate->brand =  $request->brand;
             $productUpdate->model =  $request->model;
             $productUpdate->price = $request->price;
-            foreach ($productUpdate->file as $file){
-                Storage::disk('public')->delete($file->path);
+            if($request->hasFile('file')){
+                foreach ($productUpdate->file as $file){
+                    Storage::disk('public')->delete($file->path);
+                }
             }
             $productUpdate->save();
-            $productUpdate->file()->delete();
-            foreach ($product_paths as $path){
-                $productUpdate->file()->create([
-                    'path' => $path
-                ]);
+            if($request->hasFile('file')){
+                $productUpdate->file()->delete();
+                foreach ($product_paths as $path){
+                    $productUpdate->file()->create([
+                        'path' => $path
+                    ]);
+                }
             }
             $data = [
                 'code' => 200,
@@ -80,11 +93,15 @@ class ProductService{
     public function delete($request){
         try {
             $productDelete = Product::where('id', $request->id)->with('file')->first();
-            foreach ($productDelete->file as $file){
-                Storage::disk('public')->delete($file->path);
+            if($productDelete->file){
+                foreach ($productDelete->file as $file){
+                    Storage::disk('public')->delete($file->path);
+                }
             }
             $productDelete->delete();
-            $productDelete->file()->delete();
+            if($productDelete->file){
+                $productDelete->file()->delete();
+            }
             $data = [
                 'code' => 200,
                 'status' =>  'success',
